Move stacking-context paint order into the paint layer
Appendix E ordering is paint-order logic, not DOM traversal: the walk over
fragment children, the z-index layering and the inline-pass predicate now live
in Html/Core/Paint/StackingOrder.cs. DomUtils keeps only NeedsStackingHoist and
IsStackingContextBox, which layout itself calls when computing
HasStackingHoistCandidates.

Rules are moved verbatim - no behaviour change.
