+++ src/PeachPDF/Html/Core/Utils/DomUtils.cs
@@ -618,77 +618,6 @@ namespace PeachPDF.Html.Core.Utils
             return null;
         }
 
-        // One box to paint as part of a stacking context's own layer ordering, plus the chain of DOM
-        // ancestor boxes (outer to inner, between the claiming stacking context and Box itself) that
-        // Box was hoisted past. Empty for a direct plain child - it paints via ordinary nested Paint()
-        // recursion, so its ancestors' own overflow clipping is already correctly active on the
-        // graphics clip stack from their own (still-running) Paint() calls. Non-empty for a hoisted
-        // participant - it paints via the claiming stacking context's own paint loop instead, bypassing
-        // those ancestors' Paint() calls entirely, so their overflow clipping must be reapplied
-        // explicitly (see RenderUtils.PushAncestorOverflowClips) around its own Paint() call.
-        internal readonly record struct StackingParticipant(BoxFragment Fragment, IReadOnlyList<CssBox> ClipAncestors)
-        {
-            /// <summary>The box whose style and stacking role decide where this participant paints.</summary>
-            internal CssBox Box => Fragment.Box;
-        }
-
-        /// <summary>
-        /// The participants of one fragment's own stacking context, discovered over its <i>fragment</i>
-        /// children — so a descendant that does not appear on this page is simply not found. The
-        /// ordering rules themselves are unchanged and still read the originating boxes.
-        /// </summary>
-        public static IEnumerable<StackingParticipant> FlattenStackingContext(BoxFragment fragment)
-        {
-            var box = fragment.Box;
-
-            // Plain in-flow, non-stacking-context children always paint here, nested normally - this is
-            // what keeps this box's own overflow-clip scope (pushed/popped around this same children
-            // loop in CssBox.PaintImpCore) wrapped around them, and their own further plain descendants
-            // are handled the same way, recursively, by their own subsequent Paint() call.
-            foreach (var childFragment in fragment.Children)
-            {
-                var childBox = childFragment.Box;
-
-                // ::marker boxes (inside or outside position) are always painted via one explicit
-                // Paint(g) call from CssBox.PaintImpCore/PaintListItem, not discovered generically here
-                // - both so the tagged-PDF path can wrap the marker in its own "/Lbl" structure element
-                // separately from the rest of the list item's "/LBody" content, and so an "outside"
-                // marker (which must not affect - or be discovered through - the owning list item's own
-                // stacking context, per CSS2.1 12.5.1 / CSS Lists Level 3) never gets bubbled up as if
-                // it were normal in-flow content. Yielding it here too would double-paint it.
-                if (childBox.IsMarkerPseudoElement) continue;
-
-                if (!NeedsStackingHoist(childBox))
-                {
-                    yield return new StackingParticipant(childFragment, []);
-                }
-            }
-
-            // The nearest enclosing "local ordering scope" (see IsLocalOrderingScope) is responsible for
-            // finding and ordering every out-of-flow / stacking-context-establishing descendant reachable
-            // through plain wrapper boxes AND plain floats, at any depth - a box that is neither claims
-            // nothing further here; any such descendants of its own are claimed by whichever ancestor
-            // above it actually qualifies. This is what fixes three bugs in earlier versions of this
-            // method: (1) a box that itself establishes a stacking context (e.g. position:relative;
-            // z-index, or now also opacity<1/transform) was never yielded by its own parent at all, so it
-            // and its whole subtree never painted; (2) an out-of-flow stacking-context descendant nested a
-            // few plain wrapper boxes deep was discovered "naturally" via the ordinary parent-to-child
-            // Paint() cascade before its true ancestor stacking context ever reached its own z-index
-            // layer, so it visually painted as if z-index had no effect; (3) a box that is positioned
-            // (absolute/relative/fixed/sticky) but establishes no NEW stacking context of its own
-            // (z-index:auto) never searched its own subtree either, so its own floated/positioned-without-
-            // z-index children (Appendix E's "non-positioned floats" and "positioned descendants with
-            // stack level 0") escaped all the way to the nearest TRUE stacking context ancestor instead of
-            // being ordered locally against their true DOM siblings - see IsLocalOrderingScope.
-            if (!IsLocalOrderingScope(box)) yield break;
-            if (!(box.HtmlContainer?.HasStackingHoistCandidates ?? true)) yield break;
-
-            foreach (var participant in SearchForHoistableDescendants(fragment, []))
-            {
-                yield return participant;
-            }
-        }
-
         // A box needs to escape its immediate DOM position to compete for z-order at the nearest
         // enclosing stacking context, rather than paint nested within its immediate parent: either
         // because it's out-of-flow (floated/absolute/fixed - always subject to z-ordering against its
@@ -696,88 +625,8 @@ namespace PeachPDF.Html.Core.Utils
         // stacking context (which must be ordered as one atomic unit among its true siblings, not
         // wherever it happens to sit in a plain wrapper's local scope). Internal (not private) so
         // HtmlContainerInt's HasStackingHoistCandidates computation can reuse the exact same predicate
-        // rather than duplicating it.
+        // rather than duplicating it, alongside Paint.StackingOrder, which owns the ordering walk.
         internal static bool NeedsStackingHoist(CssBox box) => box.IsOutOfFlow || IsStackingContextBox(box);
-
-        // A box claims local Appendix-E ordering responsibility for its own PLAIN FLOAT descendants
-        // (rather than deferring them to a more distant ancestor) if it is the root, a genuine stacking
-        // context, OR merely positioned (absolute/relative/fixed/sticky) regardless of z-index - matching
-        // Appendix E step 6's "positioned descendants with stack level 0 [...] painted via the same
-        // [7-step] procedure" model, under which every positioned box (not only ones with an explicit
-        // z-index) is its own atomic recursive unit for steps 3/4/5 (block/float/inline). This does NOT
-        // extend to genuine stacking-context descendants nested inside a merely-positioned (z-index:auto)
-        // box - those must keep escaping all the way to the true nearest stacking context, exactly like
-        // through a plain non-positioned wrapper, because z-index competition only happens at a REAL
-        // stacking context's own level (see the claimFloatsHere parameter on
-        // SearchForHoistableDescendants, which encodes this float-vs-stacking-context distinction - a
-        // merely-positioned box is a local ordering boundary for floats only, not for z-index).
-        //
-        // Without the float half of this, a positioned-but-z-index:auto box's own float child (Acid2's
-        // own ".eyes" - position:absolute, no z-index - containing float "#eyes-b" alongside block
-        // "#eyes-c" and inline "#eyes-a") was hoisted all the way to the true root's own stacking pass
-        // instead of being ordered locally against its true DOM siblings, painting relative to the root's
-        // entire subtree instead of interleaved correctly within ".eyes" itself.
-        private static bool IsLocalOrderingScope(CssBox box) =>
-            box.IsRoot || IsStackingContextBox(box) ||
-            box.Position is CssConstants.Absolute or CssConstants.Relative or CssConstants.Fixed or CssConstants.Sticky;
-
-        // Tunnels through plain wrapper boxes looking for content that needs to compete at `box`'s own
-        // local ordering scope. Two categories of content are hoisted here, with different stopping
-        // rules:
-        //
-        // - A genuine stacking context (IsStackingContextBox) always keeps escaping through anything
-        //   that ISN'T ITSELF a stacking context - including a merely-positioned (z-index:auto) box -
-        //   because z-index only has meaning relative to the nearest REAL stacking context. Recursion
-        //   stops at (but includes) each stacking context found; its own subtree is its own business,
-        //   resolved independently once its own Paint() call later invokes FlattenStackingContext on
-        //   itself.
-        // - A plain FLOAT only escapes as far as the nearest box that IsLocalOrderingScope (root, a
-        //   genuine stacking context, or merely positioned) - once the walk has passed through such a
-        //   box, `claimFloatsHere` flips to false for everything beneath it, since that box will find
-        //   and locally order its own floats itself (via its own later FlattenStackingContext call,
-        //   whose initial bail check now also accepts merely-positioned boxes - see IsLocalOrderingScope)
-        //   rather than this outer search claiming them too, which would both double-paint them and
-        //   order them relative to the wrong (too-distant) box's siblings.
-        //
-        // `ancestorPath` accumulates every DOM ancestor walked through along the way (both plain
-        // pass-through wrappers and hoisted-but-not-yet-fully-resolved boxes like a merely-positioned
-        // box) - each yielded participant snapshots it as its ClipAncestors, so the caller can re-apply
-        // those ancestors' own overflow clipping (which it never picks up naturally, having been hoisted
-        // past their own Paint() calls). Mutating one shared list via add-before-recurse/remove-after is
-        // safe here: the whole sequence is drained eagerly and synchronously by FlattenStackingContext's
-        // caller before anything else touches it.
-        private static IEnumerable<StackingParticipant> SearchForHoistableDescendants(
-            BoxFragment fragment, List<CssBox> ancestorPath, bool claimFloatsHere = true)
-        {
-            foreach (var childFragment in fragment.Children)
-            {
-                var childBox = childFragment.Box;
-
-                if (childBox.IsMarkerPseudoElement) continue;
-
-                var isStackingContext = IsStackingContextBox(childBox);
-                var isLocalOrderingScope = !isStackingContext && IsLocalOrderingScope(childBox);
-                var isPlainFloatToClaim = claimFloatsHere && !isStackingContext && !isLocalOrderingScope && childBox.IsOutOfFlow;
-
-                if (isStackingContext || isLocalOrderingScope || isPlainFloatToClaim)
-                {
-                    yield return new StackingParticipant(childFragment, ancestorPath.ToArray());
-                    if (isStackingContext) continue;
-                }
-
-                // Once the walk passes through a merely-positioned (non-stacking-context) box, any
-                // further floats beneath it belong to THAT box's own local claim, not this search's -
-                // only genuine stacking contexts still need to keep escaping past it.
-                var claimBeyond = isLocalOrderingScope ? false : claimFloatsHere;
-
-                ancestorPath.Add(childBox);
-                foreach (var descendant in SearchForHoistableDescendants(childFragment, ancestorPath, claimBeyond))
-                {
-                    yield return descendant;
-                }
-                ancestorPath.RemoveAt(ancestorPath.Count - 1);
-            }
-        }
 
         public static bool IsStackingContextBox(CssBox box)
         {
@@ -820,30 +669,6 @@ namespace PeachPDF.Html.Core.Utils
             }
 
             return false;
-        }
-
-        public static IEnumerable<List<StackingParticipant>> GetBoxesByLayers(IEnumerable<StackingParticipant> participants)
-        {
-            var boxesByLayer = new Dictionary<int, List<StackingParticipant>>();
-
-            foreach (var participant in participants)
-            {
-                var zIndex = 0;
-
-                if (participant.Box.ZIndex is not CssConstants.Auto)
-                {
-                    zIndex = int.Parse(participant.Box.ZIndex);
-                }
-
-                if (!boxesByLayer.ContainsKey(zIndex))
-                {
-                    boxesByLayer[zIndex] = [];
-                }
-
-                boxesByLayer[zIndex].Add(participant);
-            }
-
-            return boxesByLayer.OrderBy(x => x.Key).Select(x => x.Value);
         }
 
         public static bool IsProperTableChild(CssBox box)
